Move intent handling into app on iOS 14+, exclusively on macOS
