+++ Sources/Shared/Intents/UpdateSensorsIntentHandler.swift
@@ -1,6 +1,5 @@
 import Foundation
 import PromiseKit
-import Shared
 
 class UpdateSensorsIntentHandler: NSObject, UpdateSensorsIntentHandling {
     func handle(intent: UpdateSensorsIntent, completion: @escaping (UpdateSensorsIntentResponse) -> Void) {
